+++ ChallengePorto/src/main/java/br/com/porto/main/TipoUsuarioInsert.java
@@ -13,8 +13,8 @@ public class TipoUsuarioInsert {
 		return JOptionPane.showInputDialog(j);
 	}
 	
-	static Long id (String j) {
-		return Long.parseLong(JOptionPane.showInputDialog(j));
+	static int id (String j) {
+		return Integer.parseInt(JOptionPane.showInputDialog(j));
 	}
 	
 	public static void main(String[] args) throws ClassNotFoundException, SQLException {
